requirements.txt
aiohttp
discord
discord.py
python-dotenv
requests
py_spoo_url